Upgrading to v3.0.0
1. :fire: Fixed bug with version 2.0.2
2. Added a way to alert about unread notifications (the number displayed on the icon of the extension)
3. In the development process, a
4. :fire: Fixed bug with the Ctrl+Enter key combination handler
5. The code is rewritten more universally for different browsers, taking into account their features
6. The functional AJAX is moved to the background. Initially, there was an application architecture error (an open tab with the site was required for the expansion work). Now everything is OK :smile:
7. Added link for a quick jump to toster.ru

+++ compile.sh
@@ -9,6 +9,7 @@ FILES=("icon" "js" "css" "_locales" "resources" "manifest.json" "popup.html")
 
 $(which browserify) -t brfs ${SRC_DIR}/resources/content.js -o ${SRC_DIR}/js/content.js
 $(which browserify) -t brfs ${SRC_DIR}/resources/popup.js -o ${SRC_DIR}/js/popup.js
+$(which browserify) -t brfs ${SRC_DIR}/resources/background.js -o ${SRC_DIR}/js/background.js
 
 [ -d "${TARGET_DIR}" ] && rm -rf ${TARGET_DIR}
 [ -f "${TARGET_DIR}.zip" ] && rm -f ${TARGET_DIR}.zip
@@ -21,5 +22,6 @@ do
     cp -r ${SRC_DIR}/$file_source ${TARGET_DIR}
 done
 
-rm -f ${TARGET_DIR}/resources/content.js
-rm -f ${TARGET_DIR}/resources/popup.js
+[ -f "${TARGET_DIR}/resources/content.js" ] && rm -f ${TARGET_DIR}/resources/content.js
+[ -f "${TARGET_DIR}/resources/popup.js" ] && rm -f ${TARGET_DIR}/resources/popup.js
+[ -f "${TARGET_DIR}/resources/background.js" ] && rm -f ${TARGET_DIR}/resources/background.js
